+++ src/typey-boi/store/typey-boi.ts
@@ -1,5 +1,14 @@
 import { reducer } from 'reducers/typey-boi'
 import { createStore } from 'redux'
 import { initialState } from 'models/typey-boi'
+import { loadState, saveState } from 'store/localStorage'
 
-export const store = createStore(reducer, initialState)
+let persistedState = loadState()
+if (!persistedState) {
+  persistedState = initialState
+}
+export const store = createStore(reducer, persistedState)
+
+store.subscribe(() => {
+  saveState(store.getState())
+})